Update: Monday, April 7th
Completed the log-in PHP script and attached it the the application.  Users can now only access the main page of the application if they have an account that's registered in the database.

Created a new PHP script called GetPostingData which returns a specific piece of data from a post.

Reworked the main page to feature samples of job postings.  Three new text labels and an image used to tell the user about the job title, company hiring, a description of the job, and an image to promote the post.  All of this data is pulled from the server using the new PHP script, GetPostingData.

Notable Bug: The last job post in the database doesn't get shown.  The application acts like it should be shown, but all the information stays on the previous postings data.